Move offer.py location
